requirements.txt
Flask==3.0.2
flask-cors==4.0.0
mysql-connector-python==8.1.0
ollama==0.1.11